requirements.txt
insightface
opencv-python
numpy 
onnxruntime
psutil
python-dotenv
